Dynamic Swagger configuration: bye bye swagger.basePath.
Automatically reconfigure Swagger's basePath and apiPath from the requested
URI. This uses a lock, in case the server is accessed from different names
(e.g. directly or through a reverse proxy), so there's a (theoretical) risk
of lock contention but this is not a high-traffic endpoint so it shouldn't
be an issue in practice, and could eventually be mitigated through caching
(and alternative being to generate the Swagger description file at compile-
time, dynamically rewriting the basePath at runtime).

Also fixes links to swagger-ui resources, in case the page is accessed with
a trailing '/' in the URI (/swagger-ui.html/)

Change-Id: I801e6f1e29aded8cb9d58b06493ec3caade84355