Fix RevokeEndpoint and strengthen TokenInfo parsing
RevokeEndpoint incorrectly returned "invalid_token" errors instead of
just sending an OK response back; from RFC 7009:

   Note: invalid tokens do not cause an error response since the client
   cannot handle such an error in a reasonable way.  Moreover, the
   purpose of the revocation request, invalidating the particular token,
   is already achieved.

It otherwise failed when sending such an OK response when the token was
successfully revoked, because it passed a 'null' entity.

It also incorrectly checked that the token was attached to an existing
account: if the token exists in the datastore, we should revoke it no
matter what.

We also remove the no-cache headers: POST requests and authenticated
requests are non-cacheable by default.

Finally, TokenSerializer didn't cope for malformed base64url encoding.

h/t @schambon for reporting the issue.

Change-Id: Ic080b3ed0499484340e74444eb9018466bb402e3