Avoid unnecessary always-true check when using a refresh token
We check that the scopes being asked for are all contained in
the scopes granted to the refresh token.
But when no scope is explicitly sent, this resulted in checking
that the refresh token scopes contains all the refresh token scopes.
To avoid that, we just move the check to the case
where scopes are explicitly given in the request.

Change-Id: Ie43f3552cfd7212a4491810c39c332c7d9668ba3