Error-handling of JSON parse errors.
jackson-jaxrs-base comes with exception mappers, but given that we also
use Jackson as part of persistence (Jongo) and logging, we don't use them
as they could leak sensitive data. Instead, we wrap JacksonJsonProvider's
readFrom to catch exceptions and throw corresponding
WebApplicationExceptions.

Because we no longer use the default JacksonJsonProvider, we can also
provide our ObjectMapper directly, without the need for a ContextResolver.

Change-Id: Id287ca75dd29940f5fedbc1e2ce229f0c869ef27