Secure tokens by storing a hash of some random secret value
A random secret value is generated for each token and only ever given
to the client (combined with the token ID in the ID of the TokenInfo).
That way, if the token database is ever compromised, the attacker won't
be able to use the tokens to impersonate a user (no more than with the
database of passwords, i.e. they whould have to modify the database)

Change-Id: I088c097bf757a5e13694fa03661288576eb92764