Fix Origin handling: an origin doesn't contain a port if it's the default
One of our tests was backwards (probably a misreading of the RFC, or we
had another idea than following the RFC for the serialization; i.e. as
a simple serialization of the triple rather than creating a new class).

Browsers send "Origin: https://oasis-demo.atolcd.com" which we compared
with "https://oasis-demo.atolcd.com:443" in StrictRefererFilter.

Change-Id: Ib550a6f5d4c45e20eaa05fde4fc169cd9a91b61f