Goodbye AngularJs
Use of 'json' directive in Handlebars requires escapeHTML=true to be
XSS-safe; now that it's no longer used in the Base template, it will have
to be tested on each page that uses it. It's unlikely to ever be used
however; and the ProfilePage is scheduled for deletion so don't bother
adding a test there.

TODO: remove oasis-ui/js/** when we remove the ProfilePage (once it has
an equivalent in the OASIS Portal)

Change-Id: Ic1a9256bc1ada5ea845fe374cc1d8062f004d0aa