http.testing: commit headers after interceptors have a chance to add ones
WriterInterceptors can add headers, and they're only invoked on
ClientInvocation#writeRequestBody; so copying headers from the
ClientInvocation to the MockHttpRequest must be done afterwards to
avoid losing any header.

I verified this is also how it's implemented in the built-in
ClientHttpEngine implementations in Resteasy so it will (should)
work in a "real" environment.

Change-Id: I1614e0546783c799e80bc9a9c4f982de8d0af985