Fix redirections when query params need to contain double-%-escapes
Shitty JAX-RS UriBuilder and its crappy escaping rules!

We need to make sure that all values we pass to UriBuilder#queryParam
or WebTarget#queryParam is properly escaped, as JAX-RS won't re-escape
%-escaped sequences.

When a parameter to /a/auth contains %-escaped values and you're not
logged in, the "continue" passed to /a/login should thus contain
double-escaped sequences (%25%xx)

Maybe we should refactor the login page to read the same parameters
as /a/auth and avoid the URI-in-param-in-URI-in-param pattern?

Change-Id: I9dc268f77ec48da60e51607d6da3bec29ed1ff88