Don't use "Accept" in User-Agent fingerprinting
"Accept" can be tweaked for XMLHttpRequest, and could even depend on the
context the resource is loaded in (e.g. a document vs. a script vs. an
image). It happens that Kibana tweaks the "Accept" header to only accept
JSON or plain text, breaking authentication.

"Accept-Language" can also be tweaked, but won't be in practice in our
case, so it's safe to continue using it in the fingerprint.

Change-Id: I0bfbcd9329521c51dfb4352563814e4421ce2909