Fix UserInfoEndpoint serialization of the "locale" property
This endpoint uses the "Google OAuth Client Library for Java" for
serialization (to support JWT), which does not use our custom ObjectMapper
with its LocalizationModule and custom serializer for java.util.Locale.
That means the locale has to be converted to String on our side.

Added a quick and dirty test to make sure it works.

Change-Id: Ic7e56a1b6796511fbb85565148c0b0fe680ea53e