Update Closure Templates so it's compatible with Guava 18.0
Immutables.org requires Guava 18.0 (uses MoreObjects introduced in 18.0)
but the version of Closure Templates we used failed because of methods
that have been removed in 18.0, so we had to run with 17.0 at runtime.
Deployments were "monkey patched" in-place and we could have hacked the
build scripts to build with 18.0 but deploy 17.0, but we needed a real
fix, in case the generated code that uses MoreObjects is actually called
one day. This is that fix.

Change-Id: Id909d5b32fa8341d8e7e3cd85560f4b824d2944a