Temporarily expose the user_email_address on /apps/acl/instance
This is because a provider really needs to know the users' email address.
This is a deliberate but temporary breach into our commitment to giving
users control of their data; we'll keep this undocumented and will work
with the provider to find alternatives to even requesting the users for
their email address.

Change-Id: Iff24e6a3e1f26c465142cffe33a366ef9c2ca73f