Filter user app-instances by status, optionally include org instances
/apps/instance/user/{user_id} can now optionally include instances for
the organizations of which the user is an admin, and optionally filter
the results by instantiation status.
This allows listing all "pending instances" for a given user, whether
purchased "for personal use" or on behalf of an organization of which
he's an administrator, in a single request to the Kernel.

Change-Id: I6128b161dabd9c077a5b4c2e77123a2ccef1dee3