Make app-instance provisioning somewhat "transactional"
If there's an error when registering scopes, services, etc. while
provisioning an app-instance, set it back to the pending state and
cleanup all inserted data.
The cleanup logic is extracted from DeleteAppInstance into a new
CleanupAppInstance use-case. Changes on the AppInstance itself are
done separately so that CleanupAppInstance can be used in the
DeleteAppInstance use-case.

Change-Id: I481b17e3d160064c33c8bdf7be8115a2348417c8