Fix Service#restricted and AppInstance#redirect_uri_validation_disabled
Service#restricted wasn't deserialized because the @JsonIgnore on the
isRestricted getter was setting the whole property from being ignored.
We need to explicitly annotate the field with @JsonProperty to make it
work properly.
AppInstance#redirect_uri_validation_disabled was erroneously stored
even when 'false', because when copying AppInstance instances through
the constructor, the field was always set to a non-null value. We'll
need to update the DB manually to remove the extraneous 'false' values.

Change-Id: I2a4b5e576564b6fb17ce3ad21633a2d8fe3a372a