Quick&dirty hack to get users back to a service after sign-up
During sign-up, analyze the continueUrl to extract the client_id and
redirect_uri, lookup the service's service_uri and store it in the
AccountActivationToken. If anything goes wrong (continueUrl is not the
authorization endpoint and/or does not reference an existing service,
for instance), then just store a null value.
When the token is used to activate the account, redirect to the stored
service_uri or, if it's null, to the Portal (as before).

This is a first step towards really fixing #103. The real fix will
involve passing the authorization endpoint arguments to the login page
directly, instead of the authorization endpoint full URL as an "opaque"
(not really opaque given we're parsing it here) continueUrl. The
continueUrl will have to stay though, for those Kernel web pages that
require authentication (change password, or soon the invitation workflow,
and later the granted authorizations and token management pages).

Change-Id: Iad9765a7002e5adfdc0aff0028d1d1b78a667a1e