Make sure to use only AGPLv3-compatible dependencies
Specifically:
 * javax.servlet and javax.activation are licensed under CDDL
 * jboss-annotations-api_1.1_spec and javax.mail are dual-licensed
   under CDDL and GPLv2 with classpath exception
 * jcip-annotations are licensed under Creative Commons Attribution

All of these are replaced with Apache-licensed alternatives (most of
them from the Apache Geronimo project).

The only GPL-incompatible dependency left is dom4j, used transitively
by Swagger, and we'll get rid of it entirely at some point (sooner
rather than later now that we have a licensing issue with it).

Change-Id: I3fad02b768ef70d77db1e014d3c9f32abb29cc00