Fix Mongo indexes for organization memberships to cope for pending memberships
We need to create the index on email+accountId and make sure we store
either the email XOR the accountId (but never both at the same time,
whichever the value of 'status').
In the end, that idea of storing pending memberships inside the memberships
table was a bad one. Atomicity wasn't worth the cost.

Change-Id: Ied03a9149897e4b759be665fe1575e5d876e9d3e