Fix regression in (offline) refresh tokens
This was introduced in 9c7217111e39a7378ee8879dbca60d57220359d4 and then
made worse by e2cb2a822961abc6bcc578b6efb144d79f246cf1 that made it
unconditional.

Even with offline_access, the authorization code is a "child" of the
session (that way, we can revoke it when revoking the session, and get
the session's auth_time), but the refresh token is not (it's still a
"child" of the authorization code though, so we can eventually revoke
it in case of code reuse).

Change-Id: If3e321cc5a6340ee05da9f2c8e980e20a4be856d