Prepare ground for multiple copyright holders
Until now, all the work was being done by Atol Conseils & Développements
which also hold the copyright. Forthcoming work will be done by
Atol Conseils & Développements as a contractor, with the copyright
holder being the contractee. Also, as an open source project, it's
likely that future work will be contributed by other people or
organizations.

As such, all file headers are changed to read "The Ozwillo Kernel Authors"
as the copyright holders, and the copyright holders (currently only
Atol Conseils & Développements) are listed in an AUTHORS file.

Change-Id: I3624a6d2cb27dd1f5ec32ed7ea089816a11e3e1b