Allow authenticating with password+client cert (disabled by default)
The client certificate's Subject DN and Issuer DN must be sent by
the TLS-terminating reverse-proxy to the application through HTTP
request headers (X-SSL-Client-Subject-DN and X-SSL-Client-Issuer-DN
respectively).
If the certificate matches the current user as authenticated by his
email address and password, then the session's "level of assurance"
is "upgraded" from "low" to "substantial" (using eIDAS LoAs). This
is reflected in the 'acr' claim of the ID Tokens, which is
unconditionally sent.

Change-Id: Ic8ae89bb8d31e23bef7bed00a15ba3644a73fe40