Fix catalog not unindexing services when changed to hidden
Logic was here, but wasn't looking at the correct data, and
would never unindex a service due to a change of visibility.

Note that this is still partially flawed, as the unindexing
request outcome is ignored (only logged). Ideally, we should
unindex first then update MongoDB (so that an undexing error
would be returned to the caller; and a MongoDB error would
only produce false-negatives from the index, never false-
positives like today); that would however mean that the
index would have to be capable of checking the version for
optimistic locks; or make the update dance more complex
(e.g. get the doc from the index, then delete it, unconditionally,
then update MongoDB, and if it fails with an invalid version
then we can re-insert the document in the index).
An alternative would be to have a queue of indexation tasks,
so we could queue the deletion and have it repeated (requeued)
until it succeeds and we get into an eventually consistent state.

Change-Id: I8cba9dae619d917fa220cf5f3dc6c503471b7501

# Conflicts:
#	oasis-webapp/src/main/java/oasis/catalog/IndexingServiceRepository.java