Fix account linking to FranceConnect: SidToken wasn't updated
…which meant that the Portal couldn't retrieve the FC UserInfo,
as the access_token wasn't preserved.
The side-effect of this change is that the session is "promoted"
to "signed-in with FranceConnect", so when signing out, the
user will be redirected to FranceConnect end session endpoint,
and when unlinking their account (in the same session), they'll
be asked for their password.

Change-Id: If6eca0793e1e7523e2babd63984c0aabc3445db1