Improve user profile display in authorization endpoint
Display claims in the same order as the profile page on the
portal.
Removes the "(need your authorization)" label (this also fixes
a translation issue), and moves the "(required)" label to the
right hand side.
Rework CSS to use flex:1 on the claim value rather than a
margin-left:auto on the claim-info (renamed from info-claim),
and remove the padding-left of the row-claim (while using a
smaller padding overall.)
Lastly, fix French translation issue.

Change-Id: I9711600ec73d78613784bffb4eb1eab7d8920f75