Add an AbstractOAuthToken#isPortal to help support several portals
The flag is not (yet) initialized at that point, so to keep
backwards compatibility, isPortal falls back to testing the
client_id for the known value for the "canonical portal".
The idea is that "computing" that flag will require a Mongo
request, so we'll compute it once and store it in the tokens
for easier and faster processing.

The only uses of ClientIds.PORTAL left are in LogoutPage and
AuthorizationEndpoint (where the flag will be computed).

Change-Id: Id93f9909c2c7e70ecd8c0e1737b401c1d8cd1ae6