Manage distinct "catalogs" per portal
Each catalog entry is associated with one or more "portals".
A service is by default associated with the portal from
which its app instance has been instantiated.
The MarketSearchEndpoint filters the catalog depending on
which portal is calling.
APIs will be subsequently added to manage the association of
catalog entries to portals.

Change-Id: I0767ebdbc5b707a476fa40ae4f8eb62b97e0ff9f