Create Urls according to brand
We don't inject Urls anymore, but we create it each time we need it
with the UrlsFactory according to the brandId pass in the url or body
request.

UrlsFactory give the Urls following the portalBaseUri (fetch from db or
we take the configured portal-base-uri urls) and build myOasis related urls with
the PathUrls object which is configure in the reference.conf files.

Change-Id: I74c03c42495c44e38ea293a11b92cc8885cff6a9